# Use generics for workqueue metrics