@antoninbas 2024-09-25
I used the following benchmarks to identify the issue and validate the patch:
```go
package main

import (
	"fmt"
	"testing"

	"k8s.io/client-go/util/workqueue"
)

type sX struct {
	X int64
	Y int64
	Z int64
}

var keysStr []string
var keysPtr []*sX
var keysStruct []sX

func init() {
	for j := 0; j < 100; j++ {
		keysStr = append(keysStr, fmt.Sprintf("key-%d", j))
		keysPtr = append(keysPtr, &sX{})
		keysStruct = append(keysStruct, sX{
			X: int64(j),
		})
	}
}

func BenchmarkPointers(b *testing.B) {
	for i := 0; i < b.N; i++ {
		b.StopTimer()
		q := workqueue.NewTypedDelayingQueueWithConfig(workqueue.TypedDelayingQueueConfig[*sX]{})
		b.StartTimer()
		for j := 0; j < 100; j++ {
			q.Add(keysPtr[j])
			key, _ := q.Get()
			q.Done(key)
		}
	}
}

func BenchmarkStrings(b *testing.B) {
	for i := 0; i < b.N; i++ {
		b.StopTimer()
		q := workqueue.NewTypedDelayingQueueWithConfig(workqueue.TypedDelayingQueueConfig[string]{})
		b.StartTimer()
		for j := 0; j < 100; j++ {
			q.Add(keysStr[j])
			key, _ := q.Get()
			q.Done(key)
		}
	}
}

func BenchmarkStructs(b *testing.B) {
	for i := 0; i < b.N; i++ {
		b.StopTimer()
		q := workqueue.NewTypedDelayingQueueWithConfig(workqueue.TypedDelayingQueueConfig[sX]{})
		b.StartTimer()
		for j := 0; j < 100; j++ {
			q.Add(keysStruct[j])
			key, _ := q.Get()
			q.Done(key)
		}
	}
}
```

Results before applying the patch:
```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkPointers-16    	   63361	     19723 ns/op	    1268 B/op	     105 allocs/op
BenchmarkStrings-16     	   34248	     33197 ns/op	    7177 B/op	     407 allocs/op
BenchmarkStructs-16     	   30970	     38855 ns/op	   10500 B/op	     407 allocs/op
```

Results after applying the patch:
```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkPointers-16    	   64141	     18835 ns/op	    1239 B/op	     105 allocs/op
BenchmarkStrings-16     	   60921	     21467 ns/op	    2212 B/op	     105 allocs/op
BenchmarkStructs-16     	   47352	     25214 ns/op	    3167 B/op	     105 allocs/op
```

```
$ go version
go version go1.23.0 darwin/amd64
```

Notice that there is no extra heap allocation when casting a pointer to `interface{}`, and the results for `BenchmarkPointers` are the same before and after (this is expected: we already have a data pointer and don't need to escape to the heap). Workqueues in K8s and in K8s controllers typically use a string or a small struct.