@tnqn 2024-09-26
/ok-to-test
/lgtm

It seems this optimization could also be applied to `waitFor`, which is the last remaining struct that doesn't yet use generics in this package, aside from this one. But it could be done in a separate patch if this optimization makes sense to the workqueue maintainers.
https://github.com/kubernetes/kubernetes/blob/9770283c137e73169a8e6f94abdf3e15f1f8cc4d/staging/src/k8s.io/client-go/util/workqueue/delaying_queue.go#L175-L180