@tnqn 2024-09-26
Somehow the link checks think the updated functions are unused.