@antoninbas 2024-09-26
> Somehow the lint checks think the updated functions are unused.

Yes, I believe this is because of a bug in the `staticcheck` linter, that has been fixed, but k8s currently uses an older version of golangci-lint. I may need to open an other PR first to bump golangci-lint.