@aojea 2024-10-03
+1 for the first commit but not for the second commit.

I don't like the idea to change the code for a linter/golang error https://github.com/kubernetes/kubernetes/pull/127635#discussion_r1786781287 and I think is better to exclude the `unused` linter on that file (don't know if it works though but we should check that option)