@aojea 2024-10-03
> No problem for the requested change on my side. I will try it and update the PR if it works.

thanks,  please add a comment with the context and the references, so future us or new contributors remove the exception once the problem is fixed