@antoninbas 2024-10-03
@aojea I have updated the golangci config as we discussed, and updated the second commit accordingly.

BTW, I have also tried adding a `//nolint:unused` directive on individual methods, and it also seems to be working, assuming that the directive is added on all the methods for `noMetrics[T]` and `defaultQueueMetrics[T]` (9 in total). I do believe that the golangci config solution is better, as it gives us one central place where we can add a comment with the necessary context, and extra files can be added in the future if other parts of the K8s codebase run into the same issue. Let me know what you think.